Rename credential modules
