+++ identity_credential/src/validator/jwt_presentation_validation/error.rs
@@ -4,7 +4,7 @@
 use std::error::Error;
 use std::fmt::Display;
 
-use crate::validator::vc_jwt_validation::JwtValidationError;
+use crate::validator::jwt_credential_validation::JwtValidationError;
 
 /// Errors caused by a failure to validate a [`Presentation`](crate::presentation::Presentation).
 #[derive(Debug)]
