+++ identity_credential/src/validator/jwt_presentation_validation/jwt_presentation_validator_utils.rs
@@ -10,8 +10,8 @@ use std::str::FromStr;
 use crate::credential::Jwt;
 use crate::presentation::Presentation;
 use crate::presentation::PresentationJwtClaims;
-use crate::validator::vc_jwt_validation::JwtValidationError;
-use crate::validator::vc_jwt_validation::SignerContext;
+use crate::validator::jwt_credential_validation::JwtValidationError;
+use crate::validator::jwt_credential_validation::SignerContext;
 
 /// Utility functions for verifying JWT presentations.
 #[non_exhaustive]
