+++ identity_credential/src/validator/mod.rs
@@ -6,11 +6,11 @@
 pub use self::options::FailFast;
 pub use self::options::StatusCheck;
 pub use self::options::SubjectHolderRelationship;
-pub use self::vc_jwt_validation::*;
-pub use self::vp_jwt_validation::*;
+pub use self::jwt_credential_validation::*;
+pub use self::jwt_presentation_validation::*;
 
 mod options;
 #[cfg(test)]
 pub(crate) mod test_utils;
-mod vc_jwt_validation;
-mod vp_jwt_validation;
+mod jwt_credential_validation;
+mod jwt_presentation_validation;
