+++ identity_storage/Cargo.toml
@@ -32,6 +32,7 @@ tokio = { version = "1.29.0", default-features = false, features = ["macros", "s
 zeroize = { version = "1.6.0", default_features = false, optional = true }
 
 [dev-dependencies]
+identity_eddsa_verifier = { version = "=0.7.0-alpha.7", path = "../identity_eddsa_verifier", default-features = false, features = ["ed25519"] }
 identity_credential = { version = "=0.7.0-alpha.7", path = "../identity_credential", features = ["revocation-bitmap"] }
 once_cell = { version = "1.18", default-features = false }
 tokio = { version = "1.29.0", default-features = false, features = ["macros", "sync", "rt"] }
