+++ identity_storage/src/storage/tests/presentation_validation.rs
@@ -14,12 +14,15 @@ use identity_credential::presentation::PresentationBuilder;
 use identity_credential::validator::DecodedJwtPresentation;
 use identity_credential::validator::JwtPresentationValidationOptions;
 use identity_credential::validator::JwtPresentationValidator;
+use identity_credential::validator::JwtPresentationValidatorUtils;
 use identity_credential::validator::JwtValidationError;
 use identity_did::CoreDID;
 use identity_did::DID;
 use identity_document::document::CoreDocument;
+use identity_eddsa_verifier::EdDSAJwsVerifier;
 use identity_verification::jws::JwsAlgorithm;
 use identity_verification::MethodScope;
+use once_cell::sync::Lazy;
 
 use crate::key_storage::JwkMemStore;
 use crate::storage::tests::test_utils::generate_credential;
@@ -30,6 +33,9 @@ use crate::JwkDocumentExt;
 use crate::JwsSignatureOptions;
 
 use super::test_utils::CredentialSetup;
+
+static JWT_PRESENTATION_VALIDATOR_ED25519: Lazy<JwtPresentationValidator<EdDSAJwsVerifier>> =
+  Lazy::new(|| JwtPresentationValidator::with_signature_verifier(EdDSAJwsVerifier::default()));
 
 #[tokio::test]
 async fn test_valid_presentation() {
@@ -67,8 +73,7 @@ where
     .await
     .unwrap();
 
-  let validator: JwtPresentationValidator = JwtPresentationValidator::new();
-  let decoded_presentation: DecodedJwtPresentation<Jwt> = validator
+  let decoded_presentation: DecodedJwtPresentation<Jwt> = JWT_PRESENTATION_VALIDATOR_ED25519
     .validate::<_, Jwt, Object>(
       &presentation_jwt,
       &setup.subject_doc,
@@ -125,11 +130,10 @@ where
     .await
     .unwrap();
 
-  let holder = JwtPresentationValidator::extract_holder::<CoreDID>(&presentation_jwt).unwrap();
+  let holder = JwtPresentationValidatorUtils::extract_holder::<CoreDID>(&presentation_jwt).unwrap();
   assert_eq!(holder.to_url(), setup.subject_doc.as_ref().id().to_url());
 
-  let validator: JwtPresentationValidator = JwtPresentationValidator::new();
-  let decoded_presentation: DecodedJwtPresentation<Jwt> = validator
+  let decoded_presentation: DecodedJwtPresentation<Jwt> = JWT_PRESENTATION_VALIDATOR_ED25519
     .validate::<_, _, Object>(
       &presentation_jwt,
       &setup.subject_doc,
@@ -207,8 +211,8 @@ where
     )
     .await
     .unwrap();
-  let validator: JwtPresentationValidator = JwtPresentationValidator::new();
-  let validation_error: JwtValidationError = validator
+
+  let validation_error: JwtValidationError = JWT_PRESENTATION_VALIDATOR_ED25519
     .validate::<_, Jwt, Object>(
       &presentation_jwt,
       &setup.subject_doc,
@@ -264,8 +268,7 @@ where
     .await
     .unwrap();
 
-  let validator: JwtPresentationValidator = JwtPresentationValidator::new();
-  let validation_error: JwtValidationError = validator
+  let validation_error: JwtValidationError = JWT_PRESENTATION_VALIDATOR_ED25519
     .validate::<_, Jwt, Object>(
       &presentation_jwt,
       &setup.subject_doc,
@@ -286,7 +289,7 @@ where
   validation_options =
     validation_options.earliest_expiry_date(Timestamp::now_utc().checked_sub(Duration::days(2)).unwrap());
 
-  validator
+    JWT_PRESENTATION_VALIDATOR_ED25519
     .validate::<_, Jwt, Object>(&presentation_jwt, &setup.subject_doc, &validation_options)
     .unwrap();
 }
@@ -329,8 +332,7 @@ where
     .await
     .unwrap();
 
-  let validator: JwtPresentationValidator = JwtPresentationValidator::new();
-  let validation_error: JwtValidationError = validator
+  let validation_error: JwtValidationError = JWT_PRESENTATION_VALIDATOR_ED25519
     .validate::<_, Jwt, Object>(
       &presentation_jwt,
       &setup.subject_doc,
@@ -350,7 +352,7 @@ where
   validation_options =
     validation_options.latest_issuance_date(Timestamp::now_utc().checked_add(Duration::hours(2)).unwrap());
 
-  let validation_ok: bool = validator
+  let validation_ok: bool = JWT_PRESENTATION_VALIDATOR_ED25519
     .validate::<_, Jwt, Object>(&presentation_jwt, &setup.subject_doc, &validation_options)
     .is_ok();
   assert!(validation_ok);
@@ -390,8 +392,7 @@ where
     .await
     .unwrap();
 
-  let validator: JwtPresentationValidator = JwtPresentationValidator::new();
-  let validation_error: JwtValidationError = validator
+  let validation_error: JwtValidationError = JWT_PRESENTATION_VALIDATOR_ED25519
     .validate::<_, Jwt, Object>(
       &presentation_jwt,
       &setup.subject_doc,
